% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.semEff.R
\name{export_diagram}
\alias{export_diagram}
\title{Export semEff diagram}
\usage{
export_diagram(diagram, filename)
}
\arguments{
\item{diagram}{A diagram (output from \code{\link{plot.semEff}}).}

\item{filename}{Path to write the image.}
}
\description{
Export semEff diagram
}
