% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_stats.R
\name{inline_contrast}
\alias{inline_contrast}
\title{Inline contrast}
\usage{
inline_contrast(
  x,
  contrast = NULL,
  what = c("estimate", "ratio"),
  by = NULL,
  subset = NULL,
  at = NULL,
  digits = 3,
  label = TRUE,
  qval = FALSE
)
}
\arguments{
\item{x}{An \code{\link[emmeans:contrast]{emmeans::contrast}} table.}
}
\description{
Inline contrast
}
