% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_stats.R
\name{inline_jnt}
\alias{inline_jnt}
\title{Inline joint test}
\usage{
inline_jnt(tbl, term, at = NULL, digits = 3)
}
\arguments{
\item{tbl}{An \code{\link[emmeans:joint_tests]{emmeans::joint_tests}} table.}
}
\description{
Inline joint test
}
