% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_stats.R
\name{inline_trend}
\alias{inline_trend}
\title{Inline emtrend}
\usage{
inline_trend(tbl, term, at = NULL, label = TRUE, digits = 2, lab_lower = TRUE)
}
\arguments{
\item{tbl}{An \code{\link[emmeans:emtrends]{emmeans::emtrends}} table.}
}
\description{
Inline emtrend
}
