% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_trait.R
\name{map_trait}
\alias{map_trait}
\title{Map trait}
\usage{
map_trait(map, data, trait, lat, lon, alpha = 0.8, ...)
}
\arguments{
\item{map}{A \code{ggmap} object created by \code{\link[ggmap:get_map]{get_map}}.}

\item{data}{A \code{data.frame} containing latitude and longitude coordinates
along with trait data to plot on the map.}

\item{trait}{A string specifying the column name in \code{data} containing
the values to map.}

\item{lat}{A string specifying the column name in \code{data}
containing the latitudes.}

\item{lon}{A string specifying the column name in \code{data}
containing the longitudes.}

\item{alpha}{Transparency of dots. Argument to
\code{\link[ggplot2:geom_point]{geom_point}}. Defaults to \code{0.8}.}

\item{...}{Additional arguments to \code{\link[ggplot2:geom_point]{geom_point}}.}
}
\description{
Convenience function to plot traits or other data onto a map.
}
\details{
Relies on \code{ggmap}.
}
