% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.semEff.R
\name{plot.semEff}
\alias{plot.semEff}
\title{Plot semEff}
\usage{
\method{plot}{semEff}(mod, legend = FALSE, cols = c("blue", "red"))
}
\arguments{
\item{mod}{An \code{\link[semEff:semEff]{semEff}} model.}

\item{legend}{Logical. Whether to include the legend. Defaults to \code{FALSE}.}

\item{cols}{A vector of two colors to color the arrows. The first color is used
to denote a positive effect and the second color is used to denote a negative
effect.}
}
\description{
Plot semEff
}
\author{
Ethan Bass
}
