% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_microplate.R
\name{read_microplate}
\alias{read_microplate}
\title{Read KC4 microplate}
\usage{
read_microplate(path, long_format = TRUE)
}
\arguments{
\item{path}{Path to BioTek KC4 \code{.txt} file.}

\item{long_format}{Logical. Whether to return data in long format. Defaults
to \code{TRUE}.}
}
\description{
Read microplate data from BioTek KC4 \code{.txt} format
}
\author{
Ethan Bass
}
